requirements.txt
sounddevice>=0.4.6
vosk>=0.3.45
pyttsx3>=2.90
numpy>=1.26.4
colorama>=0.4.6

requests>=2.31.0
qdrant-client>=1.7.0
duckduckgo-search>=5.0.0
psutil>=5.9.0
python-dotenv>=1.0.0
